Fix type assignment error in multimedia analysis dialog
- Changed format_file_size parameter type from int to float
- Resolves Pylance error where float division result was assigned to int variable
- Maintains backward compatibility as integers auto-convert to floats

+++ ui/dialogs/multimedia_analysis_dialog.py
@@ -871,7 +871,7 @@ class MultimediaAnalysisDialog(QDialog):
                         '; '.join(result.get('recommendations', []))
                     ])
     
-    def format_file_size(self, size: int) -> str:
+    def format_file_size(self, size: float) -> str:
         """Format file size in human-readable format."""
         for unit in ['B', 'KB', 'MB', 'GB']:
             if size < 1024:
